link to API key creation location
The link to the home page is not the most useful when you're looking for the place to create API keys, but I get that you might want to route people through the home page for some reason so I left that link in tact (although I think it would be better to remove it).  However I added the direct link to the API keys, and added a link to the home page in preparation for possibly removing or changing the link marked "here" in the phrase "sign up for free here".   I think the only value remaining in the text surrounding that link is the fact that it expresses that the sign up is free.  I wanted to make minimal changes, but I think it would be better if that could be expressed differently.  Let me know if you'd like me to propose a slightly more complete set of changes.